+++ test/uncomplicate/neanderthal/mkl_test.clj
@@ -48,8 +48,6 @@
 (real-test/test-lapack mkl-double)
 (real-test/test-lapack mkl-float)
 
-(real-test/test-blas-sy mkl-double)
-(real-test/test-blas-sy mkl-float)
 (real-test/test-blas-sy-host mkl-double)
 (real-test/test-blas-sy-host mkl-float)